refactor: split the constructors from the Elicit class
elicit.py, which holds el.model(), el.parameter() and the other constructors,
becomes specs.py. The Elicit class moves out of __init__.py into elicit.py, so
__init__.py only re-exports the public names. tests/unit/test_elicit.py follows
its module and becomes test_specs.py.

Breaking: el.elicit is now the module of the Elicit class; the constructors
are in el.specs. el.Elicit, el.model() and the other top-level names are
unchanged.
