+++ src/elicito/types.py
@@ -11,7 +11,7 @@ import tensorflow as tf
 
 class Hyper(TypedDict):
     """
-    Typed dictionary for specification of [`hyper`][elicito.elicit.hyper]
+    Typed dictionary for specification of [`hyper`][elicito.specs.hyper]
     """
 
     name: str
@@ -68,7 +68,7 @@ class Parameter(dict[str, Any]):
 
 class QueriesDict(TypedDict, total=False):
     """
-    Typed dictionary for specification of [`queries`][elicito.elicit.Queries]
+    Typed dictionary for specification of [`queries`][elicito.specs.Queries]
     """
 
     name: str
@@ -114,7 +114,7 @@ class Target(dict[str, Any]):
 
 class ExpertDict(TypedDict, total=False):
     """
-    typed dictionary of specification of [`expert`][elicito.elicit.Expert]
+    typed dictionary of specification of [`expert`][elicito.specs.Expert]
     """
 
     ground_truth: dict[str, Any]
@@ -149,7 +149,7 @@ class Initializer(TypedDict):
 
 class Trainer(TypedDict, total=False):
     """
-    typed dictionary for specification of [`trainer`][elicito.elicit.trainer]
+    typed dictionary for specification of [`trainer`][elicito.specs.trainer]
     """
 
     method: str
@@ -191,7 +191,7 @@ class MetaSettings(TypedDict):
     """
     Typed dictionary for specification of meta settings in `Elicit`
 
-    See [`Elicit`][elicito.Elicit]
+    See [`Elicit`][elicito.elicit.Elicit]
     """
 
     dry_run: bool
